requirements.txt
ultralytics
roboflow
opencv-python
matplotlib
pandas
seaborn
scikit-learn
kagglehub
Pillow
streamlit